+++ fixtures/packages/paper-button/test.js
@@ -5,5 +5,6 @@ module.exports = {
   stdout: `[1/2] 🌀  Converting Package...
 Out directory: ${path.join(__dirname, 'generated')}
 [2/2] 🎉  Conversion Complete!`,
-  stderr: ``,
+  stderr:
+      `paper-button: package.json name is changing from "paper-button" to "@polymer/paper-button".`,
 };