+++ src/convert-workspace.ts
@@ -12,16 +12,19 @@
  * http://polymer.github.io/PATENTS.txt
  */
 
+import * as fse from 'fs-extra';
 import * as path from 'path';
 import {Analyzer, FSUrlLoader, InMemoryOverlayUrlLoader, PackageUrlResolver, ResolvedUrl} from 'polymer-analyzer';
 import {run, WorkspaceRepo} from 'polymer-workspaces';
 
+import {BowerConfig} from './bower-config';
 import {createDefaultConversionSettings, PartialConversionSettings} from './conversion-settings';
-import {generatePackageJson, readJson, writeJson} from './manifest-converter';
+import {generatePackageJson, writeJson} from './manifest-converter';
+import {YarnConfig} from './npm-config';
 import {ProjectConverter} from './project-converter';
 import {polymerFileOverrides} from './special-casing';
 import {lookupNpmPackageName, WorkspaceUrlHandler} from './urls/workspace-url-handler';
-import {exec, logRepoError, rimraf, writeFileResults} from './util';
+import {exec, logRepoError, readJsonIfExists, rimraf, writeFileResults} from './util';
 
 /**
  * Configuration options required for workspace conversions. Contains
@@ -39,15 +42,25 @@ export const GIT_STAGING_BRANCH_NAME = 'polymer-modulizer-staging';
 /**
  * For a given repo, generate a new package.json and write it to disk.
  */
-function writePackageJson(repo: WorkspaceRepo, packageVersion: string) {
+async function writePackageJson(
+    repo: WorkspaceRepo,
+    options: {version: string, flat: boolean, private: boolean}) {
   const bowerPackageName = path.basename(repo.dir);
   const bowerJsonPath = path.join(repo.dir, 'bower.json');
-  const bowerJson = readJson(bowerJsonPath);
+  const bowerJson = fse.readJSONSync(bowerJsonPath) as Partial<BowerConfig>;
   const npmPackageName =
       lookupNpmPackageName(bowerJsonPath) || bowerPackageName;
-  const packageJson =
-      generatePackageJson(bowerJson, npmPackageName, packageVersion);
-  writeJson(packageJson, repo.dir, 'package.json');
+
+  const packageJsonPath = path.join(repo.dir, 'package.json');
+  const existingPackageJson =
+      await readJsonIfExists<Partial<YarnConfig>>(packageJsonPath);
+
+  const packageJson = generatePackageJson(
+      bowerJson,
+      {name: npmPackageName, ...options},
+      undefined,
+      existingPackageJson);
+  writeJson(packageJson, packageJsonPath);
 }
 
 /**
@@ -58,7 +71,8 @@ function configureAnalyzer(options: WorkspaceConversionSettings) {
   const urlResolver = new PackageUrlResolver({packageDir: workspaceDir});
   const urlLoader = new InMemoryOverlayUrlLoader(new FSUrlLoader(workspaceDir));
   for (const [url, contents] of polymerFileOverrides) {
-    urlLoader.urlContentsMap.set(urlResolver.resolve(`polymer/${url}` as ResolvedUrl)!, contents);
+    urlLoader.urlContentsMap.set(
+        urlResolver.resolve(`polymer/${url}` as ResolvedUrl)!, contents);
   }
   return new Analyzer({
     urlLoader,
@@ -119,9 +133,12 @@ export default async function convert(options: WorkspaceConversionSettings):
   }
 
   // Generate a new package.json for each repo:
-  const packageJsonResults = await run(options.reposToConvert, async (repo) => {
-    return writePackageJson(repo, options.packageVersion);
-  });
+  const packageJsonResults = await run(
+      options.reposToConvert, async (repo) => await writePackageJson(repo, {
+                                version: options.packageVersion,
+                                flat: options.flat,
+                                private: options.private,
+                              }));
   packageJsonResults.failures.forEach(logRepoError);
 
   // Commit all changes to a staging branch for easy state resetting.